Fix crash in when the transliterated string is longer than the input
When using the icu transliterator "Russian-Latin/BGN" to transliterate
“Щ”, the result is “SHCH”, i.e. considerably longer than the input.
libtranslit doesn’t handle that case correctly and crashes.

After calling

      utrans_transUChars (icu->trans,
			  ustr, &ustrLength, ustrCapacity,
			  0, &limit,
			  &errorCode);

with “Щ” in ustr, ustrLength = limit = 1, ustrCapacity = 2,
errorCode is U_BUFFER_OVERFLOW_ERROR and both ustrLength and limit
contain the length needed for the result “SHCH”, i.e. 4.  ustr is
unchanged though. To retry one needs not only to reallocate enough to
fit the result but also to reset ustrLength and limit to the original
values.